@@ -1,45 +1,51 @@
 {
-  "name": "battleship-game",
-  "version": "0.0.0",
+  "name": "bs-game-wip",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "dev": "npx hardhat compile && vite",
-    "build": "npx hardhat compile && vite build",
-    "preview": "npx hardhat compile && vite preview",
-    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs --fix --ignore-path .gitignore",
-    "format": "prettier --write src/",
-    "type-check": "vue-tsc --noEmit -p tsconfig.app.json --composite false"
+    "build": "tsc -b && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview"
   },
   "dependencies": {
     "@metamask/detect-provider": "^2.0.0",
-    "axios": "^1.6.5",
-    "element-plus": "^2.4.2",
-    "ethers": "^5.6.4",
-    "oh-vue-icons": "^1.0.0-rc3",
-    "path": "^0.12.7",
-    "pinia": "^2.1.7",
-    "remixicon": "^3.6.0",
-    "vue": "^3.3.4",
-    "vue-router": "^4.2.5",
-    "vuedraggable": "^4.1.0"
+    "@metamask/rpc-errors": "^6.3.1",
+    "@pixi/events": "^7.4.2",
+    "@pixi/filter-advanced-bloom": "^5.1.1",
+    "@pixi/filter-bloom": "^5.1.1",
+    "@pixi/filter-glow": "^5.2.1",
+    "@pixi/graphics-extras": "^7.4.2",
+    "@pixi/react": "^7.1.2",
+    "autoprefixer": "^10.4.19",
+    "classnames": "^2.5.1",
+    "framer-motion": "^11.3.0",
+    "pixi.js": "^7.1.4",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "recharts": "^2.12.7",
+    "tailwindcss": "^3.4.4",
+    "zustand": "^4.5.4"
   },
   "devDependencies": {
-    "@metamask/rpc-errors": "^6.1.0",
-    "@nomicfoundation/hardhat-toolbox": "^2.0.2",
-    "@rushstack/eslint-patch": "^1.3.3",
-    "@tsconfig/node18": "^18.2.2",
-    "@types/node": "^18.18.5",
-    "@vitejs/plugin-vue": "^4.4.0",
-    "@vitejs/plugin-vue-jsx": "^3.0.2",
-    "@vue/tsconfig": "^0.4.0",
-    "autoprefixer": "^10.4.16",
-    "dotenv": "^16.4.1",
-    "hardhat": "^2.18.3",
-    "npm-run-all2": "^6.1.1",
-    "postcss": "^8.4.31",
-    "tailwindcss": "^3.3.5",
-    "typescript": "~5.2.0",
-    "vite": "^4.4.11",
-    "vue-tsc": "^1.8.19"
+    "@nomicfoundation/hardhat-toolbox": "^5.0.0",
+    "@trivago/prettier-plugin-sort-imports": "^4.3.0",
+    "@types/node": "^20.14.9",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@types/ws": "^8.5.11",
+    "@typescript-eslint/eslint-plugin": "^7.13.1",
+    "@typescript-eslint/parser": "^7.13.1",
+    "@vitejs/plugin-react": "^4.3.1",
+    "dotenv": "^16.4.5",
+    "eslint": "^8.57.0",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "eslint-plugin-react-refresh": "^0.4.7",
+    "postcss": "^8.4.39",
+    "prettier": "^3.3.2",
+    "sass": "^1.77.6",
+    "typescript": "^5.5.3",
+    "vite": "^5.3.1"
   }
 }
