@@ -1,86 +1,195 @@
-// // useContractStore.js
-// import { defineStore } from 'pinia'
-// import { ethers } from 'ethers'
-// import { useWalletStore } from '@/store/walletStore.js'
-// import { useMessageStore } from '@/store/messageStore.js'
-// import BattleshipJSON from '@/assets/contract/artifacts/contracts/Battleship.sol/Battleship.json'
-// import ContractAddress from '@/assets/contract/address.json'
+import { ethers, formatUnits } from 'ethers';
+import { StateCreator, create } from 'zustand';
+import { createJSONStorage, persist } from 'zustand/middleware';
 
-// export const useContractStore = defineStore('contract', {
-//   // State is a function that returns an object
-//   state: () => ({}), // Removed the contract instance from state since it's now created in getter
+import ContractAddress from '@/assets/contract/address.json';
+import BattleshipGameJson from '@/assets/contract/artifacts/contracts/BattleshipGame.sol/BattleshipGame.json';
+import getCellXY from '@/helpers/getCellXY';
+import { MOVE_FEE } from '@/lib/constants';
+import { useMessageStore } from '@/stores/messageStore';
+import { useWalletStore } from '@/stores/walletStore';
 
-//   // Getters are like computed properties for stores
-//   getters: {
-//     // The getter is now a function that returns a new contract instance whenever it's accessed
-//     getContract: (state) => {
-//       const walletStore = useWalletStore()
-//       // This will regenerate the contract instance with the current walletStore.signer
-//       return new ethers.Contract(ContractAddress.address, BattleshipJSON.abi, walletStore.signer)
-//     },
-//     getAddress: (state) => {
-//       return ContractAddress.address
-//     }
-//   },
+import { useGameStore } from './gameStore';
 
-//   // Actions can be asynchronous and are where you define methods to change state
-//   actions: {
-//     async shoot(position, isVertical) {
-//       const walletStore = useWalletStore()
-//       const messageStore = useMessageStore()
+export type ContractStore = {
+    hits: any[];
+    misses: any[];
+    graveyard: boolean[];
+    gameOver: boolean;
+    prizePool: string;
+    guessState: GuessState;
+    lastGuessCoords: number[] | null;
+    lastError: string;
+    submitGuess: (x: number, y: number) => Promise<void>;
+    resetGuessState: () => void;
+    setPrizePool: (pp: string) => void;
+    setHits: (h: any[]) => void;
+    setMisses: (m: any[]) => void;
+    setGraveyard: (g: any[]) => void;
+    setLastGuessCoords: (g: any) => void;
+};
 
-//       // Use the getter to get the contract with the current signer
-//       const contract = this.getContract
+export type GuessState =
+    | 'IDLE'
+    | 'STARTED'
+    | 'ERROR'
+    | 'TRANSACTION_SUCCESS'
+    | 'RECEIVED_RECEIPT'
+    | 'HIT'
+    | 'MISS';
 
-//       if (!walletStore.connected()) {
-//         messageStore.addMessage('[Error] wallet not connected !')
-//         return
-//       }
+export const useContractStore = create<ContractStore>(
+    persist(
+        (set, get) => ({
+            hits: [],
+            misses: [],
+            graveyard: [],
+            gameOver: false,
+            prizePool: '',
+            guessState: 'IDLE' as GuessState,
+            lastError: '',
+            lastGuessCoords: null,
 
-//       try {
-//         messageStore.addMessage(`[Shoot] Shot ${position} - ${isVertical} !`)
-//         const tx = await contract.Shoot(position)
-//         const receipt = await tx.wait()
-//         console.log(receipt)
-//         if (receipt.events[0].args.hit) {
-//           messageStore.addMessage(`[Shoot] Successfully hit XXXXXXXXX !`)
-//         } else {
-//           messageStore.addMessage(`[Shoot] Didn't hit anything but sea !`)
-//         }
-//       } catch (e) {
-//         console.log(e)
-//         messageStore.addMessage(`[Error] ${e.message}`)
-//       }
-//     },
+            submitGuess: async (x: number, y: number) => {
+                const signer = useWalletStore.getState().signer;
+                const addNewMessage = useMessageStore.getState().addNewMessage;
 
-//     async placeShip(position, isVertical) {
-//       const walletStore = useWalletStore()
-//       const messageStore = useMessageStore()
+                if (!signer) {
+                    throw new Error('No signer available.');
+                    return;
+                }
 
-//       // Use the getter to get the contract with the current signer
-//       const contract = this.getContract
+                set({ guessState: 'STARTED' });
 
-//       if (!walletStore.connected()) {
-//         messageStore.addMessage('[Error] wallet not connected !')
-//         return
-//       }
+                addNewMessage('Issuing Guess...');
+                const contract = new ethers.Contract(
+                    ContractAddress.address,
+                    BattleshipGameJson.abi,
+                    signer
+                );
+                const moveFee = ethers.parseEther(MOVE_FEE);
+                try {
+                    const submitTx = await contract.hit(x, y, {
+                        value: moveFee,
+                    });
+                    set({ guessState: 'TRANSACTION_SUCCESS' });
+                    const receipt = await submitTx.wait();
+                    addNewMessage('Issued Guess tx: ' + receipt.hash);
+                    const { allHits, allMisses, graveyard, prizePool, success, guessedCoords } =
+                        receipt.logs[0].args.toObject();
 
-//       try {
-//         messageStore.addMessage(`[PlaceShip] Ship Added ${position} - ${isVertical} !`)
-//         const tx = await contract.joinGame(position, isVertical)
-//         const receipt = await tx.wait()
-//         console.log(receipt)
-//         if (receipt.events[0].args) {
-//           messageStore.addMessage(
-//             `[PlaceShip] Successfully placed ship at: ${receipt.events[0].args.position} - IsVertical: ${receipt.events[0].args.isVertical} `
-//           )
-//         } else {
-//           messageStore.addMessage(`[PlaceShip] Failed to place ship !`)
-//         }
-//       } catch (e) {
-//         console.log(e)
-//         messageStore.addMessage(`[Error] ${e.message}`)
-//       }
-//     }
-//   }
-// })
+                    get().setHits(allHits);
+                    get().setMisses(allMisses);
+                    get().setGraveyard(graveyard);
+                    get().setPrizePool(prizePool);
+                    set({ guessState: success ? 'HIT' : 'MISS' });
+                    get().setLastGuessCoords(guessedCoords);
+                } catch (e) {
+                    const e = error as { reason?: string };
+                    set({ guessState: 'ERROR' });
+                    if (e?.reason) {
+                        addNewMessage('Failed to issue Guess - ' + e?.reason + ' ...', 'ERROR');
+                        set({ lastError: 'Failed to issue Guess - ' + e?.reason });
+
+                        if (e?.reason === 'Cell already hit') {
+                            useGameStore.getState().addUnknownCell(x, y);
+                        }
+                    } else {
+                        addNewMessage(
+                            'Failed to issue Guess - unexpected error occurred, check the console logs...',
+                            'ERROR'
+                        );
+                        set({ lastError: 'Failed to issue Guess - unexpected error occurred' });
+                        throw new Error(e);
+                    }
+                }
+            },
+
+            resetGuessState: () =>
+                set({
+                    guessState: 'IDLE',
+                }),
+
+            setGraveyard: (latestGraveyard: any[]) => {
+                const addNewMessage = useMessageStore.getState().addNewMessage;
+
+                const graveyardHasUpdated =
+                    get().graveyard.length !== latestGraveyard.length ||
+                    get().graveyard.some((value, index) => value !== latestGraveyard[index]);
+
+                if (graveyardHasUpdated) {
+                    set({ graveyard: latestGraveyard });
+                    addNewMessage('Graveyard info updated.');
+                }
+            },
+
+            setMisses: (latestMisses: any[]) => {
+                const currentMisses = useGameStore.getState().missedCells;
+                const addNewMessage = useMessageStore.getState().addNewMessage;
+                const missesHaveUpdated = latestMisses.length !== currentMisses.length;
+
+                if (missesHaveUpdated) {
+                    const missedCells = latestMisses.map((entry) => {
+                        const [x, y] = getCellXY(parseInt(entry[0]), parseInt(entry[1]));
+                        return {
+                            col: parseInt(entry[0]),
+                            row: parseInt(entry[1]),
+                            x,
+                            y,
+                            state: 'MISSED',
+                        };
+                    });
+
+                    set({ misses: latestMisses });
+                    useGameStore.setState({ missedCells });
+                    useGameStore.getState().setRevealedCells(latestMisses, 'MISS');
+                    useGameStore.getState().clearUnknownCells();
+                    addNewMessage('Missed. Shot failed to find target.');
+                }
+            },
+
+            //TODO: Given the similarity of the methods here might be worth combining with the above.
+            setHits: (latestHits: any[]) => {
+                const currentHits = useGameStore.getState().hitCells;
+                const addNewMessage = useMessageStore.getState().addNewMessage;
+                const hitsHaveUpdated = latestHits.length !== currentHits.length;
+
+                if (hitsHaveUpdated) {
+                    const hitCells = latestHits.map((entry) => {
+                        const [x, y] = getCellXY(parseInt(entry[0]), parseInt(entry[1]));
+                        return {
+                            col: parseInt(entry[0]),
+                            row: parseInt(entry[1]),
+                            x,
+                            y,
+                            state: 'MISSED',
+                        };
+                    });
+
+                    set({ hits: latestHits });
+                    useGameStore.setState({ hitCells });
+                    useGameStore.getState().setRevealedCells(latestHits, 'HIT');
+                    addNewMessage('DIRECT HIT. Shot successfully found target.', 'SUCCESS');
+                }
+            },
+
+            setPrizePool: (prizePool: string) => {
+                const addNewMessage = useMessageStore.getState().addNewMessage;
+
+                addNewMessage(
+                    `[BattleshipGame Contract] Prize pool at: ${formatUnits(prizePool, 'ether')} ETH`
+                );
+                set({ prizePool: formatUnits(prizePool, 'ether') });
+            },
+
+            setLastGuessCoords: (coords: any) => {
+                set({ lastGuessCoords: [parseInt(coords[0]), parseInt(coords[1])] });
+            },
+        }),
+        {
+            name: 'contract-storage',
+            storage: createJSONStorage(() => localStorage),
+            partialize: (state) => ({ prizePool: state.prizePool, graveyard: state.graveyard }),
+        }
+    ) as StateCreator<ContractStore, [], []>
+);
